[#114] Fix "/bin/sh: 1: 0: not found" error
This problem has been noticed on Ubuntu 20.04.1 LTS.
It looks that /bin/sh is a symlink to dash (see below) while bash
interpreter is expected:

lrwxrwxrwx 1 root root 4 paź 12 20:35 /bin/sh -> dash

Use SHELL=/bin/bash explicitly in makbet.mk to indicate that bash
will be used as a shell interpreter.  It is of course expected that
/bin/bash exists in the system (so far there is no additional test to
check it).